LC-4sum [Runtime: N/A; Memory: N/A]

+++ two-pointers/4sum.java
@@ -10,7 +10,7 @@ class Solution {
                 int right=nums.length-1;
                 while(left<right)
                 {
-                    long sum=nums[i]+nums[j]+nums[left]+nums[right];
+                    long sum=(long)nums[i]+nums[j]+nums[left]+nums[right];
                     if(sum==target)
                     {
                       List<Integer> list=new ArrayList<>();
